Setup React and folder struc.
